+++ source/renderer/app/api/api.js
@@ -81,7 +81,7 @@ import {
 } from '../ipc/cardano.ipc';
 import patchAdaApi from './utils/patchAdaApi';
 import { getLegacyWalletId, utcStringToDate } from './utils';
-import { Logger } from '../utils/logging';
+import { logger } from '../utils/logging';
 import {
   unscrambleMnemonics,
   scrambleMnemonics,
@@ -221,7 +221,7 @@ export default class AdaApi {
   }
 
   getWallets = async (): Promise<Array<Wallet>> => {
-    Logger.debug('AdaApi::getWallets called');
+    logger.debug('AdaApi::getWallets called');
     try {
       const wallets: AdaWallets = isIncentivizedTestnet
         ? await getWallets(this.config)
@@ -229,7 +229,7 @@ export default class AdaApi {
       const legacyWallets: LegacyAdaWallets = await getLegacyWallets(
         this.config
       );
-      Logger.debug('AdaApi::getWallets success', { wallets, legacyWallets });
+      logger.debug('AdaApi::getWallets success', { wallets, legacyWallets });
 
       map(legacyWallets, legacyAdaWallet => {
         const extraLegacyWalletProps = {
@@ -249,13 +249,13 @@ export default class AdaApi {
 
       return wallets.map(_createWalletFromServerData);
     } catch (error) {
-      Logger.error('AdaApi::getWallets error', { error });
+      logger.error('AdaApi::getWallets error', { error });
       throw new GenericApiError();
     }
   };
 
   getWallet = async (request: GetWalletRequest): Promise<Wallet> => {
-    Logger.debug('AdaApi::getWallet called', {
+    logger.debug('AdaApi::getWallet called', {
       parameters: filterLogData(request),
     });
     try {
@@ -282,10 +282,10 @@ export default class AdaApi {
       } else {
         wallet = await getWallet(this.config, { walletId });
       }
-      Logger.debug('AdaApi::getWallet success', { wallet });
+      logger.debug('AdaApi::getWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::getWallet error', { error });
+      logger.error('AdaApi::getWallet error', { error });
       throw new GenericApiError();
     }
   };
@@ -293,7 +293,7 @@ export default class AdaApi {
   getAddresses = async (
     request: GetAddressesRequest
   ): Promise<Array<WalletAddress>> => {
-    Logger.debug('AdaApi::getAddresses called', {
+    logger.debug('AdaApi::getAddresses called', {
       parameters: filterLogData(request),
     });
     const { walletId, queryParams, isLegacy } = request;
@@ -302,10 +302,10 @@ export default class AdaApi {
       if (!isLegacy) {
         response = await getAddresses(this.config, walletId, queryParams);
       }
-      Logger.debug('AdaApi::getAddresses success', { addresses: response });
+      logger.debug('AdaApi::getAddresses success', { addresses: response });
       return response.map(_createAddressFromServerData);
     } catch (error) {
-      Logger.error('AdaApi::getAddresses error', { error });
+      logger.error('AdaApi::getAddresses error', { error });
       throw new GenericApiError();
     }
   };
@@ -313,7 +313,7 @@ export default class AdaApi {
   getTransactions = async (
     request: GetTransactionsRequest
   ): Promise<GetTransactionsResponse> => {
-    Logger.debug('AdaApi::searchHistory called', { parameters: request });
+    logger.debug('AdaApi::searchHistory called', { parameters: request });
     const { walletId, order, fromDate, toDate, isLegacy } = request;
 
     const params = Object.assign(
@@ -345,7 +345,7 @@ export default class AdaApi {
         resolve({ transactions, total: response.length })
       );
     } catch (error) {
-      Logger.error('AdaApi::searchHistory error', { error });
+      logger.error('AdaApi::searchHistory error', { error });
       throw new GenericApiError();
     }
 
@@ -354,7 +354,7 @@ export default class AdaApi {
     // const requestStats = Object.assign({}, request, {
     //   cachedTransactions: request.cachedTransactions.length,
     // });
-    //  Logger.debug('AdaApi::searchHistory called', { parameters: requestStats });
+    //  logger.debug('AdaApi::searchHistory called', { parameters: requestStats });
     // const requestTimestamp = moment();
     // const params = {
     //   wallet_id: walletId,
@@ -489,19 +489,19 @@ export default class AdaApi {
     //       .duration(moment().diff(requestTimestamp))
     //       .as('milliseconds'),
     //   };
-    //   Logger.debug(
+    //   logger.debug(
     //     `AdaApi::searchHistory success: ${total} transactions loaded`,
     //     { responseStats }
     //   );
     //   return new Promise(resolve => resolve({ transactions, total }));
     // } catch (error) {
-    //   Logger.error('AdaApi::searchHistory error', { error });
+    //   logger.error('AdaApi::searchHistory error', { error });
     //   throw new GenericApiError();
     // }
   };
 
   createWallet = async (request: CreateWalletRequest): Promise<Wallet> => {
-    Logger.debug('AdaApi::createWallet called', {
+    logger.debug('AdaApi::createWallet called', {
       parameters: filterLogData(request),
     });
     const { name, mnemonic, spendingPassword } = request;
@@ -516,7 +516,7 @@ export default class AdaApi {
         wallet = await createWallet(this.config, {
           walletInitData,
         });
-        Logger.debug('AdaApi::createWallet (Shelley) success', { wallet });
+        logger.debug('AdaApi::createWallet (Shelley) success', { wallet });
       } else {
         const legacyWallet: LegacyAdaWallet = await restoreByronWallet(
           this.config,
@@ -536,17 +536,17 @@ export default class AdaApi {
           ...legacyWallet,
           ...extraLegacyWalletProps,
         };
-        Logger.debug('AdaApi::createWallet (Byron) success', { wallet });
+        logger.debug('AdaApi::createWallet (Byron) success', { wallet });
       }
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::createWallet error', { error });
+      logger.error('AdaApi::createWallet error', { error });
       throw new GenericApiError();
     }
   };
 
   deleteWallet = async (request: DeleteWalletRequest): Promise<boolean> => {
-    Logger.debug('AdaApi::deleteWallet called', {
+    logger.debug('AdaApi::deleteWallet called', {
       parameters: filterLogData(request),
     });
     try {
@@ -557,10 +557,10 @@ export default class AdaApi {
       } else {
         response = await deleteWallet(this.config, { walletId });
       }
-      Logger.debug('AdaApi::deleteWallet success', { response });
+      logger.debug('AdaApi::deleteWallet success', { response });
       return true;
     } catch (error) {
-      Logger.error('AdaApi::deleteWallet error', { error });
+      logger.error('AdaApi::deleteWallet error', { error });
       throw new GenericApiError();
     }
   };
@@ -568,7 +568,7 @@ export default class AdaApi {
   createTransaction = async (
     request: CreateTransactionRequest
   ): Promise<WalletTransaction> => {
-    Logger.debug('AdaApi::createTransaction called', {
+    logger.debug('AdaApi::createTransaction called', {
       parameters: filterLogData(request),
     });
     const { walletId, address, amount, passphrase, isLegacy } = request;
@@ -600,13 +600,13 @@ export default class AdaApi {
         });
       }
 
-      Logger.debug('AdaApi::createTransaction success', {
+      logger.debug('AdaApi::createTransaction success', {
         transaction: response,
       });
 
       return _createTransactionFromServerData(response);
     } catch (error) {
-      Logger.error('AdaApi::createTransaction error', { error });
+      logger.error('AdaApi::createTransaction error', { error });
       if (error.code === 'output_is_redeem') {
         throw new NotAllowedToSendMoneyToRedeemAddressError();
       }
@@ -633,7 +633,7 @@ export default class AdaApi {
   calculateTransactionFee = async (
     request: GetTransactionFeeRequest
   ): Promise<BigNumber> => {
-    Logger.debug('AdaApi::calculateTransactionFee called', {
+    logger.debug('AdaApi::calculateTransactionFee called', {
       parameters: filterLogData(request),
     });
     const {
@@ -682,12 +682,12 @@ export default class AdaApi {
         throw new NotEnoughFundsForTransactionFeesError();
       }
 
-      Logger.debug('AdaApi::calculateTransactionFee success', {
+      logger.debug('AdaApi::calculateTransactionFee success', {
         transactionFee: response,
       });
       return fee;
     } catch (error) {
-      Logger.error('AdaApi::calculateTransactionFee error', { error });
+      logger.error('AdaApi::calculateTransactionFee error', { error });
       if (error.name === 'NotEnoughFundsForTransactionFeesError') {
         throw error;
       } else if (error.code === 'not_enough_money') {
@@ -718,7 +718,7 @@ export default class AdaApi {
   deleteTransaction = async (
     request: DeleteTransactionRequest
   ): Promise<void> => {
-    Logger.debug('AdaApi::deleteTransaction called', { parameters: request });
+    logger.debug('AdaApi::deleteTransaction called', { parameters: request });
     const { walletId, transactionId, isLegacy } = request;
     try {
       let response;
@@ -733,9 +733,9 @@ export default class AdaApi {
           transactionId,
         });
       }
-      Logger.debug('AdaApi::deleteTransaction success', response);
+      logger.debug('AdaApi::deleteTransaction success', response);
     } catch (error) {
-      Logger.error('AdaApi::deleteTransaction error', { error });
+      logger.error('AdaApi::deleteTransaction error', { error });
       // In this particular call we don't need to handle the error in the UI
       // The only reason transaction canceling would fail is if the transaction
       // is no longer pending - in which case there is nothign we can do.
@@ -746,7 +746,7 @@ export default class AdaApi {
     mnemonic.split(' ').length === ADA_CERTIFICATE_MNEMONIC_LENGTH;
 
   getWalletRecoveryPhrase(): Promise<Array<string>> {
-    Logger.debug('AdaApi::getWalletRecoveryPhrase called');
+    logger.debug('AdaApi::getWalletRecoveryPhrase called');
     try {
       const response: Promise<Array<string>> = new Promise(resolve =>
         resolve(
@@ -757,24 +757,24 @@ export default class AdaApi {
           )
         )
       );
-      Logger.debug('AdaApi::getWalletRecoveryPhrase success');
+      logger.debug('AdaApi::getWalletRecoveryPhrase success');
       return response;
     } catch (error) {
-      Logger.error('AdaApi::getWalletRecoveryPhrase error', { error });
+      logger.error('AdaApi::getWalletRecoveryPhrase error', { error });
       throw new GenericApiError();
     }
   }
 
   getWalletCertificateAdditionalMnemonics(): Promise<Array<string>> {
-    Logger.debug('AdaApi::getWalletCertificateAdditionalMnemonics called');
+    logger.debug('AdaApi::getWalletCertificateAdditionalMnemonics called');
     try {
       const response: Promise<Array<string>> = new Promise(resolve =>
         resolve(generateAdditionalMnemonics())
       );
-      Logger.debug('AdaApi::getWalletCertificateAdditionalMnemonics success');
+      logger.debug('AdaApi::getWalletCertificateAdditionalMnemonics success');
       return response;
     } catch (error) {
-      Logger.error('AdaApi::getWalletCertificateAdditionalMnemonics error', {
+      logger.error('AdaApi::getWalletCertificateAdditionalMnemonics error', {
         error,
       });
       throw new GenericApiError();
@@ -784,16 +784,16 @@ export default class AdaApi {
   getWalletCertificateRecoveryPhrase(
     request: GetWalletCertificateRecoveryPhraseRequest
   ): Promise<Array<string>> {
-    Logger.debug('AdaApi::getWalletCertificateRecoveryPhrase called');
+    logger.debug('AdaApi::getWalletCertificateRecoveryPhrase called');
     const { passphrase, input: scrambledInput } = request;
     try {
       const response: Promise<Array<string>> = new Promise(resolve =>
         resolve(scrambleMnemonics({ passphrase, scrambledInput }))
       );
-      Logger.debug('AdaApi::getWalletCertificateRecoveryPhrase success');
+      logger.debug('AdaApi::getWalletCertificateRecoveryPhrase success');
       return response;
     } catch (error) {
-      Logger.error('AdaApi::getWalletCertificateRecoveryPhrase error', {
+      logger.error('AdaApi::getWalletCertificateRecoveryPhrase error', {
         error,
       });
       throw new GenericApiError();
@@ -803,14 +803,14 @@ export default class AdaApi {
   getWalletRecoveryPhraseFromCertificate(
     request: GetWalletRecoveryPhraseFromCertificateRequest
   ): Promise<Array<string>> {
-    Logger.debug('AdaApi::getWalletRecoveryPhraseFromCertificate called');
+    logger.debug('AdaApi::getWalletRecoveryPhraseFromCertificate called');
     const { passphrase, scrambledInput } = request;
     try {
       const response = unscrambleMnemonics({ passphrase, scrambledInput });
-      Logger.debug('AdaApi::getWalletRecoveryPhraseFromCertificate success');
+      logger.debug('AdaApi::getWalletRecoveryPhraseFromCertificate success');
       return Promise.resolve(response);
     } catch (error) {
-      Logger.error('AdaApi::getWalletRecoveryPhraseFromCertificate error', {
+      logger.error('AdaApi::getWalletRecoveryPhraseFromCertificate error', {
         error,
       });
       return Promise.reject(new InvalidMnemonicError());
@@ -818,7 +818,7 @@ export default class AdaApi {
   }
 
   restoreWallet = async (request: RestoreWalletRequest): Promise<Wallet> => {
-    Logger.debug('AdaApi::restoreWallet called', {
+    logger.debug('AdaApi::restoreWallet called', {
       parameters: filterLogData(request),
     });
     const { recoveryPhrase, walletName, spendingPassword } = request;
@@ -831,10 +831,10 @@ export default class AdaApi {
       const wallet: AdaWallet = await restoreWallet(this.config, {
         walletInitData,
       });
-      Logger.debug('AdaApi::restoreWallet success', { wallet });
+      logger.debug('AdaApi::restoreWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::restoreWallet error', { error });
+      logger.error('AdaApi::restoreWallet error', { error });
       if (error.code === 'wallet_already_exists') {
         throw new WalletAlreadyRestoredError();
       }
@@ -856,7 +856,7 @@ export default class AdaApi {
   restoreLegacyWallet = async (
     request: RestoreLegacyWalletRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::restoreLegacyWallet called', {
+    logger.debug('AdaApi::restoreLegacyWallet called', {
       parameters: filterLogData(request),
     });
     const { recoveryPhrase, walletName, spendingPassword } = request;
@@ -883,10 +883,10 @@ export default class AdaApi {
         ...legacyWallet,
         ...extraLegacyWalletProps,
       };
-      Logger.debug('AdaApi::restoreLegacyWallet success', { wallet });
+      logger.debug('AdaApi::restoreLegacyWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::restoreLegacyWallet error', { error });
+      logger.error('AdaApi::restoreLegacyWallet error', { error });
       if (error.code === 'wallet_already_exists') {
         throw new WalletAlreadyRestoredError();
       }
@@ -908,7 +908,7 @@ export default class AdaApi {
   restoreByronRandomWallet = async (
     request: RestoreLegacyWalletRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::restoreByronRandomWallet called', {
+    logger.debug('AdaApi::restoreByronRandomWallet called', {
       parameters: filterLogData(request),
     });
     const { recoveryPhrase, walletName, spendingPassword } = request;
@@ -937,10 +937,10 @@ export default class AdaApi {
         ...legacyWallet,
         ...extraLegacyWalletProps,
       };
-      Logger.debug('AdaApi::restoreByronRandomWallet success', { wallet });
+      logger.debug('AdaApi::restoreByronRandomWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::restoreByronRandomWallet error', { error });
+      logger.error('AdaApi::restoreByronRandomWallet error', { error });
       if (error.code === 'wallet_already_exists') {
         throw new WalletAlreadyRestoredError();
       }
@@ -962,7 +962,7 @@ export default class AdaApi {
   restoreByronIcarusWallet = async (
     request: RestoreLegacyWalletRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::restoreByronIcarusWallet called', {
+    logger.debug('AdaApi::restoreByronIcarusWallet called', {
       parameters: filterLogData(request),
     });
     const { recoveryPhrase, walletName, spendingPassword } = request;
@@ -991,10 +991,10 @@ export default class AdaApi {
         ...legacyWallet,
         ...extraLegacyWalletProps,
       };
-      Logger.debug('AdaApi::restoreByronIcarusWallet success', { wallet });
+      logger.debug('AdaApi::restoreByronIcarusWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::restoreByronIcarusWallet error', { error });
+      logger.error('AdaApi::restoreByronIcarusWallet error', { error });
       if (error.code === 'wallet_already_exists') {
         throw new WalletAlreadyRestoredError();
       }
@@ -1016,7 +1016,7 @@ export default class AdaApi {
   restoreByronTrezorWallet = async (
     request: RestoreLegacyWalletRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::restoreByronTrezorWallet called', {
+    logger.debug('AdaApi::restoreByronTrezorWallet called', {
       parameters: filterLogData(request),
     });
     const { recoveryPhrase, walletName, spendingPassword } = request;
@@ -1045,10 +1045,10 @@ export default class AdaApi {
         ...legacyWallet,
         ...extraLegacyWalletProps,
       };
-      Logger.debug('AdaApi::restoreByronTrezorWallet success', { wallet });
+      logger.debug('AdaApi::restoreByronTrezorWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::restoreByronTrezorWallet error', { error });
+      logger.error('AdaApi::restoreByronTrezorWallet error', { error });
       if (error.code === 'wallet_already_exists') {
         throw new WalletAlreadyRestoredError();
       }
@@ -1070,7 +1070,7 @@ export default class AdaApi {
   restoreByronLedgerWallet = async (
     request: RestoreLegacyWalletRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::restoreByronLedgerWallet called', {
+    logger.debug('AdaApi::restoreByronLedgerWallet called', {
       parameters: filterLogData(request),
     });
     const { recoveryPhrase, walletName, spendingPassword } = request;
@@ -1099,10 +1099,10 @@ export default class AdaApi {
         ...legacyWallet,
         ...extraLegacyWalletProps,
       };
-      Logger.debug('AdaApi::restoreByronLedgerWallet success', { wallet });
+      logger.debug('AdaApi::restoreByronLedgerWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::restoreByronLedgerWallet error', { error });
+      logger.error('AdaApi::restoreByronLedgerWallet error', { error });
       if (error.code === 'wallet_already_exists') {
         throw new WalletAlreadyRestoredError();
       }
@@ -1124,7 +1124,7 @@ export default class AdaApi {
   importWalletFromKey = async (
     request: ImportWalletFromKeyRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::importWalletFromKey called', {
+    logger.debug('AdaApi::importWalletFromKey called', {
       parameters: filterLogData(request),
     });
     const { filePath, spendingPassword } = request;
@@ -1133,10 +1133,10 @@ export default class AdaApi {
         filePath,
         spendingPassword: spendingPassword || '',
       });
-      Logger.debug('AdaApi::importWalletFromKey success', { importedWallet });
+      logger.debug('AdaApi::importWalletFromKey success', { importedWallet });
       return _createWalletFromServerData(importedWallet);
     } catch (error) {
-      Logger.error('AdaApi::importWalletFromKey error', { error });
+      logger.error('AdaApi::importWalletFromKey error', { error });
       if (error.message === 'WalletAlreadyExists') {
         throw new WalletAlreadyImportedError();
       }
@@ -1147,7 +1147,7 @@ export default class AdaApi {
   importWalletFromFile = async (
     request: ImportWalletFromFileRequest
   ): Promise<Wallet> => {
-    Logger.debug('AdaApi::importWalletFromFile called', {
+    logger.debug('AdaApi::importWalletFromFile called', {
       parameters: filterLogData(request),
     });
     const { filePath, spendingPassword } = request;
@@ -1163,10 +1163,10 @@ export default class AdaApi {
             spendingPassword,
           })
         : await importWalletAsJSON(this.config, filePath);
-      Logger.debug('AdaApi::importWalletFromFile success', { importedWallet });
+      logger.debug('AdaApi::importWalletFromFile success', { importedWallet });
       return _createWalletFromServerData(importedWallet);
     } catch (error) {
-      Logger.error('AdaApi::importWalletFromFile error', { error });
+      logger.error('AdaApi::importWalletFromFile error', { error });
       if (error.message === 'WalletAlreadyExists') {
         throw new WalletAlreadyImportedError();
       }
@@ -1175,18 +1175,18 @@ export default class AdaApi {
   };
 
   nextUpdate = async (): Promise<NodeSoftware | null> => {
-    Logger.debug('AdaApi::nextUpdate called');
+    logger.debug('AdaApi::nextUpdate called');
 
     /* TODO: Re-enable when API is available
     try {
       const nodeUpdate = await getNextNodeUpdate(this.config);
       if (nodeUpdate && nodeUpdate.version) {
-        Logger.debug('AdaApi::nextUpdate success', { nodeUpdate });
+        logger.debug('AdaApi::nextUpdate success', { nodeUpdate });
         return nodeUpdate;
       }
-      Logger.debug('AdaApi::nextUpdate success: No Update Available');
+      logger.debug('AdaApi::nextUpdate success: No Update Available');
     } catch (error) {
-      Logger.error('AdaApi::nextUpdate error', { error });
+      logger.error('AdaApi::nextUpdate error', { error });
       throw new GenericApiError();
     }
     */
@@ -1195,30 +1195,30 @@ export default class AdaApi {
   };
 
   postponeUpdate = async (): Promise<void> => {
-    Logger.debug('AdaApi::postponeUpdate called');
+    logger.debug('AdaApi::postponeUpdate called');
     try {
       const response: Promise<any> = await postponeNodeUpdate(this.config);
-      Logger.debug('AdaApi::postponeUpdate success', { response });
+      logger.debug('AdaApi::postponeUpdate success', { response });
     } catch (error) {
-      Logger.error('AdaApi::postponeUpdate error', { error });
+      logger.error('AdaApi::postponeUpdate error', { error });
       throw new GenericApiError();
     }
   };
 
   applyUpdate = async (): Promise<void> => {
-    Logger.debug('AdaApi::applyUpdate called');
+    logger.debug('AdaApi::applyUpdate called');
     try {
       await awaitUpdateChannel.send();
       const response: Promise<any> = await applyNodeUpdate(this.config);
-      Logger.debug('AdaApi::applyUpdate success', { response });
+      logger.debug('AdaApi::applyUpdate success', { response });
     } catch (error) {
-      Logger.error('AdaApi::applyUpdate error', { error });
+      logger.error('AdaApi::applyUpdate error', { error });
       throw new GenericApiError();
     }
   };
 
   updateWallet = async (request: UpdateWalletRequest): Promise<Wallet> => {
-    Logger.debug('AdaApi::updateWallet called', {
+    logger.debug('AdaApi::updateWallet called', {
       parameters: filterLogData(request),
     });
     const { walletId, name } = request;
@@ -1227,10 +1227,10 @@ export default class AdaApi {
         walletId,
         name,
       });
-      Logger.debug('AdaApi::updateWallet success', { wallet });
+      logger.debug('AdaApi::updateWallet success', { wallet });
       return _createWalletFromServerData(wallet);
     } catch (error) {
-      Logger.error('AdaApi::updateWallet error', { error });
+      logger.error('AdaApi::updateWallet error', { error });
       throw new GenericApiError();
     }
   };
@@ -1238,7 +1238,7 @@ export default class AdaApi {
   updateSpendingPassword = async (
     request: UpdateSpendingPasswordRequest
   ): Promise<boolean> => {
-    Logger.debug('AdaApi::updateSpendingPassword called', {
+    logger.debug('AdaApi::updateSpendingPassword called', {
       parameters: filterLogData(request),
     });
     const { walletId, oldPassword, newPassword } = request;
@@ -1248,10 +1248,10 @@ export default class AdaApi {
         oldPassword,
         newPassword,
       });
-      Logger.debug('AdaApi::updateSpendingPassword success');
+      logger.debug('AdaApi::updateSpendingPassword success');
       return true;
     } catch (error) {
-      Logger.error('AdaApi::updateSpendingPassword error', { error });
+      logger.error('AdaApi::updateSpendingPassword error', { error });
       const errorCode = get(error, 'code', '');
       if (
         errorCode === 'wrong_encryption_passphrase' ||
@@ -1267,7 +1267,7 @@ export default class AdaApi {
   quitStakePool = async (
     request: QuitStakePoolRequest
   ): Promise<Transaction> => {
-    Logger.debug('AdaApi::quitStakePool called', {
+    logger.debug('AdaApi::quitStakePool called', {
       parameters: filterLogData(request),
     });
     const { walletId, passphrase } = request;
@@ -1276,10 +1276,10 @@ export default class AdaApi {
         walletId,
         passphrase,
       });
-      Logger.debug('AdaApi::quitStakePool success', { result });
+      logger.debug('AdaApi::quitStakePool success', { result });
       return result;
     } catch (error) {
-      Logger.error('AdaApi::quitStakePool error', { error });
+      logger.error('AdaApi::quitStakePool error', { error });
       const errorCode = get(error, 'code', '');
       if (
         errorCode === 'wrong_encryption_passphrase' ||
@@ -1296,7 +1296,7 @@ export default class AdaApi {
     request: ExportWalletToFileRequest
   ): Promise<[]> => {
     const { walletId, filePath } = request;
-    Logger.debug('AdaApi::exportWalletToFile called', {
+    logger.debug('AdaApi::exportWalletToFile called', {
       parameters: filterLogData(request),
     });
     try {
@@ -1304,10 +1304,10 @@ export default class AdaApi {
         walletId,
         filePath,
       });
-      Logger.debug('AdaApi::exportWalletToFile success', { response });
+      logger.debug('AdaApi::exportWalletToFile success', { response });
       return response;
     } catch (error) {
-      Logger.error('AdaApi::exportWalletToFile error', { error });
+      logger.error('AdaApi::exportWalletToFile error', { error });
       throw new GenericApiError();
     }
   };
@@ -1316,17 +1316,17 @@ export default class AdaApi {
     request: GetWalletUtxosRequest
   ): Promise<WalletUtxos> => {
     const { walletId } = request;
-    Logger.debug('AdaApi::getWalletUtxos called', {
+    logger.debug('AdaApi::getWalletUtxos called', {
       parameters: filterLogData(request),
     });
     try {
       const response: Promise<WalletUtxos> = await getWalletUtxos(this.config, {
         walletId,
       });
-      Logger.debug('AdaApi::getWalletUtxos success', { response });
+      logger.debug('AdaApi::getWalletUtxos success', { response });
       return response;
     } catch (error) {
-      Logger.error('AdaApi::getWalletUtxos error', { error });
+      logger.error('AdaApi::getWalletUtxos error', { error });
       throw new GenericApiError();
     }
   };
@@ -1335,7 +1335,7 @@ export default class AdaApi {
     request: GetWalletIdAndBalanceRequest
   ): Promise<WalletIdAndBalance> => {
     const { recoveryPhrase, getBalance } = request;
-    Logger.debug('AdaApi::getWalletIdAndBalance called', {
+    logger.debug('AdaApi::getWalletIdAndBalance called', {
       parameters: { getBalance },
     });
     try {
@@ -1346,7 +1346,7 @@ export default class AdaApi {
           getBalance,
         }
       );
-      Logger.debug('AdaApi::getWalletIdAndBalance success', { response });
+      logger.debug('AdaApi::getWalletIdAndBalance success', { response });
       const { walletId, balance } = response;
       return {
         walletId,
@@ -1356,7 +1356,7 @@ export default class AdaApi {
             : null,
       };
     } catch (error) {
-      Logger.error('AdaApi::getWalletIdAndBalance error', { error });
+      logger.error('AdaApi::getWalletIdAndBalance error', { error });
       throw new GenericApiError();
     }
   };
@@ -1365,7 +1365,7 @@ export default class AdaApi {
     request: ForceWalletResyncRequest
   ): Promise<void> => {
     await wait(FORCED_WALLET_RESYNC_WAIT); // API request throttling
-    Logger.debug('AdaApi::forceWalletResync called', { parameters: request });
+    logger.debug('AdaApi::forceWalletResync called', { parameters: request });
     try {
       const { walletId, isLegacy } = request;
       let response;
@@ -1374,9 +1374,9 @@ export default class AdaApi {
       } else {
         response = await forceWalletResync(this.config, { walletId });
       }
-      Logger.debug('AdaApi::forceWalletResync success', { response });
+      logger.debug('AdaApi::forceWalletResync success', { response });
     } catch (error) {
-      Logger.error('AdaApi::forceWalletResync error', { error });
+      logger.error('AdaApi::forceWalletResync error', { error });
       throw new GenericApiError();
     }
   };
@@ -1385,7 +1385,7 @@ export default class AdaApi {
     request: TransferFundsCalculateFeeRequest
   ): Promise<BigNumber> => {
     const { sourceWalletId } = request;
-    Logger.debug('AdaApi::transferFundsCalculateFee called', {
+    logger.debug('AdaApi::transferFundsCalculateFee called', {
       parameters: { sourceWalletId },
     });
     try {
@@ -1395,10 +1395,10 @@ export default class AdaApi {
           sourceWalletId,
         }
       );
-      Logger.debug('AdaApi::transferFundsCalculateFee success', { response });
+      logger.debug('AdaApi::transferFundsCalculateFee success', { response });
       return _createMigrationFeeFromServerData(response);
     } catch (error) {
-      Logger.error('AdaApi::transferFundsCalculateFee error', { error });
+      logger.error('AdaApi::transferFundsCalculateFee error', { error });
       throw new GenericApiError();
     }
   };
@@ -1407,7 +1407,7 @@ export default class AdaApi {
     request: TransferFundsRequest
   ): Promise<TransferFundsResponse> => {
     const { sourceWalletId, targetWalletId, passphrase } = request;
-    Logger.debug('AdaApi::transferFunds called', {
+    logger.debug('AdaApi::transferFunds called', {
       parameters: { sourceWalletId, targetWalletId },
     });
 
@@ -1417,10 +1417,10 @@ export default class AdaApi {
         targetWalletId,
         passphrase,
       });
-      Logger.debug('AdaApi::transferFunds success', { response });
+      logger.debug('AdaApi::transferFunds success', { response });
       return response;
     } catch (error) {
-      Logger.error('AdaApi::transferFunds error', { error });
+      logger.error('AdaApi::transferFunds error', { error });
       if (
         error.code === 'wrong_encryption_passphrase' ||
         (error.code === 'bad_request' &&
@@ -1433,32 +1433,34 @@ export default class AdaApi {
   };
 
   getStakePools = async (): Promise<Array<StakePool>> => {
-    Logger.debug('AdaApi::getStakePools called');
+    logger.debug('AdaApi::getStakePools called');
     try {
       const response: AdaApiStakePools = await getStakePools(this.config);
       const stakePools = response
         .filter(({ metadata }: AdaApiStakePool) => metadata !== undefined)
         .map(_createStakePoolFromServerData);
-      Logger.debug('AdaApi::getStakePools success', {
+      logger.debug('AdaApi::getStakePools success', {
         stakePoolsTotal: response.length,
         stakePoolsWithMetadata: stakePools.length,
       });
       return stakePools;
     } catch (error) {
-      Logger.error('AdaApi::getStakePools error', { error });
+      logger.error('AdaApi::getStakePools error', { error });
       throw new GenericApiError();
     }
   };
 
   testReset = async (): Promise<void> => {
-    Logger.debug('AdaApi::testReset called');
+    logger.debug('AdaApi::testReset called');
     try {
-      const wallets: AdaWallets = await getWallets(this.config);
-      Promise.all(
-        wallets.map(wallet =>
-          deleteWallet(this.config, { walletId: wallet.id })
-        )
-      );
+      if (isIncentivizedTestnet) {
+        const wallets: AdaWallets = await getWallets(this.config);
+        Promise.all(
+          wallets.map(wallet =>
+            deleteWallet(this.config, { walletId: wallet.id })
+          )
+        );
+      }
       const legacyWallets: LegacyAdaWallets = await getLegacyWallets(
         this.config
       );
@@ -1467,20 +1469,20 @@ export default class AdaApi {
           deleteLegacyWallet(this.config, { walletId: wallet.id })
         )
       );
-      Logger.debug('AdaApi::testReset success');
+      logger.debug('AdaApi::testReset success');
     } catch (error) {
-      Logger.error('AdaApi::testReset error', { error });
+      logger.error('AdaApi::testReset error', { error });
       throw new GenericApiError();
     }
   };
 
   getNetworkInfo = async (): Promise<GetNetworkInfoResponse> => {
-    Logger.debug('AdaApi::getNetworkInfo called');
+    logger.debug('AdaApi::getNetworkInfo called');
     try {
       const networkInfo: NetworkInfoResponse = await getNetworkInfo(
         this.config
       );
-      Logger.debug('AdaApi::getNetworkInfo success', { networkInfo });
+      logger.debug('AdaApi::getNetworkInfo success', { networkInfo });
       /* eslint-disable-next-line camelcase */
       const { sync_progress, node_tip, network_tip, next_epoch } = networkInfo;
       const syncProgress =
@@ -1512,7 +1514,7 @@ export default class AdaApi {
         },
       };
     } catch (error) {
-      Logger.error('AdaApi::getNetworkInfo error', { error });
+      logger.error('AdaApi::getNetworkInfo error', { error });
       // @API TODO - Inspect this implementation once TLS support is implemented on the BE
       if (error.code === TlsCertificateNotValidError.API_ERROR) {
         throw new TlsCertificateNotValidError();
@@ -1522,7 +1524,7 @@ export default class AdaApi {
   };
 
   getLatestAppVersion = async (): Promise<GetLatestAppVersionResponse> => {
-    Logger.debug('AdaApi::getLatestAppVersion called');
+    logger.debug('AdaApi::getLatestAppVersion called');
     try {
       const { isWindows, platform } = global.environment;
       const latestAppVersionInfo: LatestAppVersionInfoResponse = await getLatestAppVersion();
@@ -1546,20 +1548,20 @@ export default class AdaApi {
         applicationVersionPath,
         null
       );
-      Logger.debug('AdaApi::getLatestAppVersion success', {
+      logger.debug('AdaApi::getLatestAppVersion success', {
         latestAppVersion,
         latestAppVersionInfo,
         applicationVersion,
       });
       return { latestAppVersion, applicationVersion };
     } catch (error) {
-      Logger.error('AdaApi::getLatestAppVersion error', { error });
+      logger.error('AdaApi::getLatestAppVersion error', { error });
       throw new GenericApiError();
     }
   };
 
   getNews = async (): Promise<GetNewsResponse> => {
-    Logger.debug('AdaApi::getNews called');
+    logger.debug('AdaApi::getNews called');
 
     // Fetch news json
     let rawNews: string;
@@ -1568,7 +1570,7 @@ export default class AdaApi {
       rawNews = await getNews();
       news = JSON.parse(rawNews);
     } catch (error) {
-      Logger.error('AdaApi::getNews error', { error });
+      logger.error('AdaApi::getNews error', { error });
       throw new Error('Unable to fetch news');
     }
 
@@ -1579,7 +1581,7 @@ export default class AdaApi {
       newsHash = await getSHA256HexForString(rawNews);
       expectedNewsHash = await getNewsHash(news.updatedAt);
     } catch (error) {
-      Logger.error('AdaApi::getNews (hash) error', { error });
+      logger.error('AdaApi::getNews (hash) error', { error });
       throw new Error('Unable to fetch news hash');
     }
 
@@ -1587,7 +1589,7 @@ export default class AdaApi {
       throw new Error('Newsfeed could not be verified');
     }
 
-    Logger.debug('AdaApi::getNews success', {
+    logger.debug('AdaApi::getNews success', {
       updatedAt: news.updatedAt,
       items: news.items.length,
     });
@@ -1597,18 +1599,18 @@ export default class AdaApi {
   calculateDelegationFee = async (
     request: GetDelegationFeeRequest
   ): Promise<BigNumber> => {
-    Logger.debug('AdaApi::calculateDelegationFee called', {
+    logger.debug('AdaApi::calculateDelegationFee called', {
       parameters: filterLogData(request),
     });
     try {
       const response: DelegationFee = await getDelegationFee(this.config, {
         walletId: request.walletId,
       });
-      Logger.debug('AdaApi::calculateDelegationFee success', { response });
+      logger.debug('AdaApi::calculateDelegationFee success', { response });
       const delegationFee = _createDelegationFeeFromServerData(response);
       return delegationFee;
     } catch (error) {
-      Logger.error('AdaApi::calculateDelegationFee error', { error });
+      logger.error('AdaApi::calculateDelegationFee error', { error });
       throw new GenericApiError();
     }
   };
@@ -1616,7 +1618,7 @@ export default class AdaApi {
   joinStakePool = async (
     request: JoinStakePoolRequest
   ): Promise<Transaction> => {
-    Logger.debug('AdaApi::joinStakePool called', {
+    logger.debug('AdaApi::joinStakePool called', {
       parameters: filterLogData(request),
     });
     const { walletId, stakePoolId, passphrase } = request;
@@ -1626,12 +1628,12 @@ export default class AdaApi {
         stakePoolId,
         passphrase,
       });
-      Logger.debug('AdaApi::joinStakePool success', {
+      logger.debug('AdaApi::joinStakePool success', {
         stakePool: response,
       });
       return response;
     } catch (error) {
-      Logger.error('AdaApi::joinStakePool error', { error });
+      logger.error('AdaApi::joinStakePool error', { error });
       if (
         error.code === 'wrong_encryption_passphrase' ||
         (error.code === 'bad_request' &&